requirement.txt
torch
torchvision
torchaudio
matplotlib
seaborn
numpy
scikit-learn
scikit-image
tqdm
pillow